Fix syntax error for 0.13.x version of `purs`
In 0.13.0, the parser was changed so it wouldn't allow strings with
invalid Unicode. `\ffff` is now a syntax error. To get around this
issue, we move it to the FFI.

This isn't the only way to fix this problem, but it should still allow
the same behavior of the test without changing the spirit of the test.
If there's a different way to solve this problem that we'd prefer, we
should do that instead.